Break out JourneyStatusChip to standalone component
