+++ src/components/journeys/JourneyStatusChip.tsx
@@ -1,0 +1,45 @@
+import { useIntl } from 'react-intl';
+import { Chip, makeStyles } from '@material-ui/core';
+
+import { ZetkinJourneyInstance } from 'types/zetkin';
+
+const useStyles = makeStyles((theme) => ({
+  closedChip: {
+    backgroundColor: theme.palette.error.main,
+    color: 'white',
+    fontWeight: 'bold',
+    marginRight: '1rem',
+  },
+  openChip: {
+    backgroundColor: theme.palette.success.main,
+    color: 'white',
+    fontWeight: 'bold',
+    marginRight: '1rem',
+  },
+}));
+
+interface JourneyStatusChipProps {
+  instance: Pick<ZetkinJourneyInstance, 'closed'>;
+}
+
+const JourneyStatusChip: React.FC<JourneyStatusChipProps> = ({ instance }) => {
+  const intl = useIntl();
+  const classes = useStyles();
+  return !instance.closed ? (
+    <Chip
+      className={classes.openChip}
+      label={intl.formatMessage({
+        id: 'layout.organize.journeys.statusOpen',
+      })}
+    />
+  ) : (
+    <Chip
+      className={classes.closedChip}
+      label={intl.formatMessage({
+        id: 'layout.organize.journeys.statusClosed',
+      })}
+    />
+  );
+};
+
+export default JourneyStatusChip;
